requirements.txt
appnope==0.1.0
decorator==4.0.9
Django==1.9.4
django-forms-bootstrap==3.0.1
djangorestframework==3.3.3
gnureadline==6.3.3
ipython==4.1.2
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
ptyprocess==0.5.1
simplegeneric==0.8.1
traitlets==4.2.1
